refactor(format_message): and add tests
